fix(NanoSpec): Add static_cast for double to long conversions (C4244) in 8 more files
- Fixed 26 C4244 warnings across dialog and setting files
- SpecifyCoordinate2PointDlg.cpp: 12 fixes
- SpecifyCoordinateDlg.cpp: 12 fixes
- SampleSettingDlg.cpp: 2 fixes
- Added static_cast<long>() for MICROMETRE calculations

+++ SRC/NanoSpec/SampleSettingDlg.cpp
@@ -262,8 +262,8 @@ BOOL CSampleSettingDlg::CheckData()
 	long lSampleYEdgeMin;
 	long lSampleYEdgeMax;
 
-	coordLeftTop.lX = m_StageConfig.StageMaxCoord.dLeftX + m_StageConfig.Edge.dX;
-	coordLeftTop.lY = m_StageConfig.StageMaxCoord.dUpY - m_StageConfig.Edge.dY;
+	coordLeftTop.lX = static_cast<long>(m_StageConfig.StageMaxCoord.dLeftX + m_StageConfig.Edge.dX);
+	coordLeftTop.lY = static_cast<long>(m_StageConfig.StageMaxCoord.dUpY - m_StageConfig.Edge.dY);
 	coordRightBottom.lX = m_StageConfig.StageMaxCoord.dRightX - m_StageConfig.Edge.dX;
 	coordRightBottom.lY = m_StageConfig.StageMaxCoord.dDownY + m_StageConfig.Edge.dY;
