+++ SRC/NanoSpec/SpecifyCoordinate2PointDlg.cpp
@@ -168,32 +168,32 @@ BOOL CSpecifyCoordinate2PointDlg::CheckData()
 
 		switch ( m_StageConfig.OrgPos ) {
 		case 1: // 中心
-			coordLeftTop.lX = m_dSampleSizeX / 2 * -1;
-			coordLeftTop.lY = m_dSampleSizeY / 2;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX / 2 * -1);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY / 2);
 			coordRightBottom.lX = m_dSampleSizeX / 2;
 			coordRightBottom.lY = m_dSampleSizeY / 2 * -1;
 			break;
 		case 2: // 左上
-			coordLeftTop.lX = 0;
-			coordLeftTop.lY = 0;
+			coordLeftTop.lX = static_cast<long>(0);
+			coordLeftTop.lY = static_cast<long>(0);
 			coordRightBottom.lX = m_dSampleSizeX;
 			coordRightBottom.lY = m_dSampleSizeY * -1;
 			break;
 		case 3: // 左下
-			coordLeftTop.lX = 0;
-			coordLeftTop.lY = m_dSampleSizeY;
+			coordLeftTop.lX = static_cast<long>(0);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY);
 			coordRightBottom.lX = m_dSampleSizeX;
 			coordRightBottom.lY = 0;
 			break;
 		case 4: // 右下
-			coordLeftTop.lX = m_dSampleSizeX * -1;
-			coordLeftTop.lY = m_dSampleSizeY;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX * -1);
+			coordLeftTop.lY = static_cast<long>(m_dSampleSizeY);
 			coordRightBottom.lX = 0;
 			coordRightBottom.lY = 0;
 			break;
 		case 5: // 右上
-			coordLeftTop.lX = m_dSampleSizeX * -1;
-			coordLeftTop.lY = 0;
+			coordLeftTop.lX = static_cast<long>(m_dSampleSizeX * -1);
+			coordLeftTop.lY = static_cast<long>(0);
 			coordRightBottom.lX = 0;
 			coordRightBottom.lY = m_dSampleSizeY * -1;
 			break;
@@ -207,8 +207,8 @@ BOOL CSpecifyCoordinate2PointDlg::CheckData()
 	///// 指定座標が、ステージ稼動範囲であるか
 	case SPECIFY_COORDINATE_STAGE:
 	default:
-		coordLeftTop.lX = m_StageConfig.StageMaxCoord.dLeftX;
-		coordLeftTop.lY = m_StageConfig.StageMaxCoord.dUpY;
+		coordLeftTop.lX = static_cast<long>(m_StageConfig.StageMaxCoord.dLeftX);
+		coordLeftTop.lY = static_cast<long>(m_StageConfig.StageMaxCoord.dUpY);
 		coordRightBottom.lX = m_StageConfig.StageMaxCoord.dRightX;
 		coordRightBottom.lY = m_StageConfig.StageMaxCoord.dDownY;
